Change various 64 bit typedefs over to 32 bit.
This also removes all of the changes that were needed for those
typedefs to be 64 bit.

+++ system/include/libc/sys/resource.h
@@ -16,7 +16,7 @@ extern "C" {
 
 #include <bits/alltypes.h>
 
-typedef unsigned long long rlim_t;
+typedef unsigned long rlim_t;
 
 struct rlimit
 {
